Fix bit rot after a short hiatus
The composer file was out of date and terribly out of date at that. As such
I have updated it to match the new Composer format. I have also updated the
composer.json with the latest formatting and bumped the version number to
5.3.2 as required by the Symfony Components due to a critical bug in PHP
5.3.2.

UIn addition the compile script was moved to a 'bin' folder to make it
clearer that this is a binary.