Add Vagrant configuration
With this commit I have added a vagrant configuration for Cilex. The
puppet configuration is specialized for the Ubuntu Precise Pangolin
base box but it works fine for me.

The only issue that is unsolved is that the phpunit binary is not
executable and thus fails if the executable bit is not added.

In addition to adding a vagrant configuration I have also included
PHPUnit in the composer.json and updated the dependencies.